base: init-install-efi: only copy startup.nsh if available
Make sure to check if startup.nsh is available first before trying to
copy to the /boot partition, to avoid aborting the install process when
the file is not available via the wic image.

(cherry picked from commit 27aa5cc8ad59be7f3eee18ff863a797e2844117e)

+++ meta-lmp-base/recipes-core/initrdscripts/initramfs-module-install-efi/init-install-efi.sh
@@ -270,7 +270,9 @@ configfile /boot/loader/grub.cfg
 EOF
 
 # Make sure startup.nsh is also available at the boot partition
-cp /run/media/$1/startup.nsh /boot
+if [ -f /run/media/$1/startup.nsh ]; then
+    cp /run/media/$1/startup.nsh /boot
+fi
 
 # Set default EFI boot entry
 BOOTLABEL="LmP"
